+++ src/include/Shader.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#include "Renderer.h"
+#include <string>
+#include <iostream>
+#include <fstream>
+#include <string>
+#include <sstream>
+
+struct ShaderProgramSource{
+    std::string VertexSource;
+    std::string FragmentSource;
+};
+
+class Shader{
+private:
+    std::string m_FilePath;
+    unsigned int m_RendererID;
+public:
+    Shader(const std::string& filepath);
+    ~Shader();
+
+    void Bind() const;
+    void Unbind() const;
+
+    void SetUniform4f(const std::string& name, float v0, float v1, float v2, float v3);
+private:
+    static ShaderProgramSource ParseShader(const std::string& filepath);
+    unsigned int CompileShader(unsigned int type, const std::string& source);
+    unsigned int CreateShader(const std::string& vertexShader, const std::string& fragmentShader);
+    int GetUniformLocation(const std::string& name);
+};
