+++ src/main.cpp
@@ -11,80 +11,10 @@
 #include "VertexBuffer.h"
 #include "IndexBuffer.h"
 #include "VertexArray.h"
+#include "Shader.h"
 
 #define WINDOW_HEIGHT 800
 #define WINDOW_WIDTH 600
-
-struct ShaderProgramSource{
-    std::string VertexSource;
-    std::string FragmentSource;
-};
-
-static ShaderProgramSource ParseShader(const std::string& filepath){
-    enum class ShaderType{
-        NONE = 1, VERTEX = 0, FRAGMENT = 1
-    };
-
-    std::ifstream stream(filepath);
-    std::string line;
-    std::stringstream ss[2];
-
-    ShaderType type = ShaderType::NONE;
-    while(getline(stream, line)){
-        if(line.find("#shader") != std::string::npos){
-            if(line.find("vertex") != std::string::npos){
-                type = ShaderType::VERTEX;
-            }
-            else if(line.find("fragment") != std::string::npos){
-                type = ShaderType::FRAGMENT;
-            }
-        }
-        else{
-            ss[(int) type] << line << "\n";
-        }
-    }
-
-    return { ss[0].str(), ss[1].str() };
-}
-
-
-static unsigned int CompileShader(unsigned int type, const std::string& source){
-    unsigned int id = glCreateShader(type);
-    const char* src = source.c_str();
-    GLCall(glShaderSource(id, 1, &src, nullptr));
-    GLCall(glCompileShader(id));
-
-    int result;
-    GLCall(glGetShaderiv(id, GL_COMPILE_STATUS, &result));
-    if(result == GL_FALSE){
-        int length;
-        GLCall(glGetShaderiv(id, GL_INFO_LOG_LENGTH, &length));
-        char* message = (char*)alloca(length * sizeof(char));
-        GLCall(glGetShaderInfoLog(id, length, &length, message));
-        std::cout << "Failed to compile " << (type == GL_VERTEX_SHADER ? "vertex" : "fragment") << " shader" << std::endl;
-        std::cout << message << std::endl;
-        GLCall(glDeleteShader(id));
-        return 0;
-    }
-    return id;
-}
-
-static int CreateShader(const std::string& vertexShader, const std::string& fragmentShader){
-    unsigned int program = glCreateProgram();
-    unsigned int vs = CompileShader(GL_VERTEX_SHADER, vertexShader);
-    unsigned int fs = CompileShader(GL_FRAGMENT_SHADER, fragmentShader);
-
-    GLCall(glAttachShader(program, vs));
-    GLCall(glAttachShader(program, fs));
-
-    GLCall(glLinkProgram(program));
-    GLCall(glValidateProgram(program));
-
-    GLCall(glDeleteShader(vs));
-    GLCall(glDeleteShader(fs));
-
-    return program;
-}
 
 void changeColor(
     float* red, 
@@ -159,10 +89,6 @@ int main(){
             2, 3, 0
         };
 
-        unsigned int vertexArrayObject;
-        GLCall(glGenVertexArrays(1, &vertexArrayObject));
-        GLCall(glBindVertexArray(vertexArrayObject));
-
         VertexArray va;
         VertexBuffer vb(positions, 4 * 2 * sizeof(float));
         VertexBufferLayout layout;
@@ -171,13 +97,9 @@ int main(){
 
         IndexBuffer ib(indices, 6);
 
-        ShaderProgramSource source = ParseShader("../src/res/shaders/Basic.shader");
-
-        unsigned int shader = CreateShader(source.VertexSource, source.FragmentSource);
-        glUseProgram(shader);
-
-        GLCall(int location = glGetUniformLocation(shader, "u_Color"));
-        ASSERT(location != -1);
+        Shader shader("../src/res/shaders/Basic.shader");
+        shader.Bind();
+        shader.SetUniform4f("u_Color", 0, 0, 0, 1.0f);
 
         float red = 0.5f;
         float blue = 0.2f;
@@ -190,17 +112,17 @@ int main(){
         int counter = 0;
         int counterLimit = 10;
         
-        GLCall(glBindVertexArray(0));
-        GLCall(glUseProgram(0));
-        GLCall(glBindBuffer(GL_ARRAY_BUFFER, 0));
-        GLCall(glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0));
+        va.Unbind();
+        vb.Unbind();
+        ib.Unbind();
+        shader.Unbind();
 
         while(!glfwWindowShouldClose(window)){
             GLCall(glClear(GL_COLOR_BUFFER_BIT));
 
-            GLCall(glUseProgram(shader));
+            shader.Bind();
             
-            GLCall(glUniform4f(location, red, green, blue, 1.0f));
+            shader.SetUniform4f("u_Color", red, green, blue, 1.0f);
             va.Bind();
             ib.Bind();
 
@@ -214,8 +136,6 @@ int main(){
             GLCall(glfwSwapBuffers(window));
             GLCall(glfwPollEvents());
         }
-        GLCall(glDeleteProgram(shader));
-        
     }
     glfwTerminate();
     return 0;
